feature mail confirmation DONE

+++ src/lib/Services/Auth.js
@@ -1,6 +1,6 @@
 import Publisher from '../Connexion/Publisher'
 import jwt_decode from 'jwt-decode'
-
+import VueSimpleAlert from 'vue-simple-alert'
 class Auth extends Publisher {
   constructor() {
     super('http://localhost:81/', {
@@ -12,19 +12,29 @@ class Auth extends Publisher {
   async login(data) {
     const response = await this.post('login_check', data);
     if (response?.token) {
-      localStorage.setItem('token', response.token);
-      this.token = response.token;
-
       var decoded = jwt_decode(response.token);
+      console.log(decoded);
+      if (decoded.mail_confirmed == true ||decoded.mail_confirmed == null) {
+        localStorage.setItem('token', response.token);
 
-      this.id = decoded.id;
-      this.roles = decoded.roles;
-      return response;
+        this.token = response.token;
+        this.id = decoded.id;
+        this.roles = decoded.roles;
+
+        return response;
+      }
+      else {
+        VueSimpleAlert.fire({
+          title: 'Email non confirmé',
+          text: 'Veuillez confirmer votre email avant de vous connecter !',
+          type: 'warning',
+        })
+        return false;
+      }
     }
     else {
       return false;
     }
-
   }
 
   async refresh(data) {
@@ -33,7 +43,7 @@ class Auth extends Publisher {
   }
 
   async me(properties = ['id', 'email', 'roles', 'username', 'ownGames', 'wishGames', 'receivedExchanges', 'sendExchanges']) {
-  
+
     if (this.getId()) {
       const response = await this.get(this.formatEndPoint('users', this.getId(), properties));
       return response;
@@ -52,7 +62,7 @@ class Auth extends Publisher {
     }
     return null;
   }
-  
+
   getRoles() {
     if (localStorage.getItem('token')) {
       var decoded = jwt_decode(localStorage.getItem('token'));
@@ -60,7 +70,7 @@ class Auth extends Publisher {
     }
     return null;
   }
-  
+
   getToken() {
     return localStorage.getItem('token');
   }
